Add BoosterManager for cell index debug
