+++ Scripts/BrickBlast/Gameplay/Managers/BoosterManager.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 6e591838abae40bfbdf5915ab2686356
+guid: d325e1ecc3904c808e49f9fca189e235
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+  userData:
+  assetBundleName:
+  assetBundleVariant:
